Fix the adaptation of the slider value to step size.
If step: 20, max: 50 and from/to is equal to 40-50.

Old behavior: if from/to get values 40-49 they still equal to 40
else if get value 50 they will jump to 50, because
the lower step point is 40 and upper step point is 60.

Current behavior: if from/to get values that is less than 45 they still
equal to 40 else if get values that is greater than or equal to 45
they will jump to 50.